requirements.txt
pandas
PyMySQL
sqlalchemy
feedparser
pyyaml
uuid
nltk
python-twitter
schedule
google-cloud-logging